move not-changing form components to corresponding new folders (#208)
