+++ src/components/form/index.ts
@@ -1,6 +1,6 @@
-export * from './Form'
+export * from './Checkbox'
 export * from './FieldSet'
-export * from './AutoSubmit'
+export * from './Form'
 export * from './Input'
 export * from './InputError'
 export * from './InputLabel'
@@ -10,9 +10,4 @@ export * from './SelectMonth'
 export * from './SelectYear'
 export * from './TextArea'
 export * from './ToggleSwitch'
-export * from './Checkbox'
-export * from './FormSubmitFeedback'
-export * from './useCustomInputCss'
-export * from './useDelayedFormSubmitState'
 export * from './types'
-export * from './FormError'
