+++ src/components/react-hook-form/FormSubmitFeedback.stories.mdx
@@ -12,7 +12,7 @@ import { Form } from './Form'
 import { FormSubmitFeedback } from './FormSubmitFeedback'
 
 <Meta
-  title="Components/Form/FormSubmitFeedback"
+  title="Components/React Hook Form/FormSubmitFeedback"
   component={FormSubmitFeedback}
   decorators={[
     (Story) => {
