+++ src/components/react-hook-form/async/SelectItemInput.tsx
@@ -4,10 +4,10 @@ import classNames from 'classnames'
 import { ReactNode, useMemo, useRef } from 'react'
 import { useInternationalization, useTheme } from '../../../framework'
 import { Mode } from '../../types'
-import { InputError } from '../InputError'
-import { InputLabel } from '../InputLabel'
-import { FormVariant } from '../types'
-import { useCustomInputCss } from '../useCustomInputCss'
+import { InputError } from '../../form/InputError'
+import { InputLabel } from '../../form/InputLabel'
+import { FormVariant } from '../../form/types'
+import { useCustomInputCss } from '../../form/useCustomInputCss'
 import { replacePlaceholderColorWithTextColor } from './replacePlaceholderColorWithTextColor'
 
 export type SelectItemInputProps<ItemType> = {
