+++ src/components/react-hook-form/async/index.ts
@@ -1,4 +1,4 @@
 export * from './replacePlaceholderColorWithTextColor'
-export * from './SelectItem'
+export * from './SelectItemFormField'
 export * from './SelectItemDialogWithSearch'
 export * from './SelectItemInput'
