+++ src/components/react-hook-form/async/replacePlaceholderColorWithTextColor.ts
