+++ src/components/section/SectionFooter/SectionFooterWithSubmit.tsx
@@ -5,7 +5,7 @@ import { SubmitButton } from '../../button'
 import {
   FormSubmitFeedback,
   FormSubmitFeedbackProps,
-} from '../../form/FormSubmitFeedback'
+} from '../../react-hook-form/FormSubmitFeedback'
 import { SectionFooterArea } from './SectionFooterArea'
 
 export type SectionFooterWithSubmitProps = {
