+++ src/index.ts
@@ -5,7 +5,7 @@ export * from './components/button'
 export * from './components/content'
 export * from './components/dialog'
 export * from './components/form'
-export * from './components/form/async'
+export * from './components/react-hook-form/async'
 export * from './components/header'
 export * from './components/link'
 export * from './components/loading'
